+++ lib/views/settings.dart
@@ -4,6 +4,7 @@ import 'package:cccc/services/firebase_auth.dart';
 import 'package:cccc/services/logger_init.dart';
 import 'package:cccc/styles/text_styles.dart';
 import 'package:cccc/views/linked_accounts.dart';
+import 'package:cccc/views/privacy_and_security_settings.dart';
 import 'package:cccc/widgets/show_adaptive_alert_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -32,6 +33,10 @@ class Settings extends ConsumerWidget {
           ListTile(
             onTap: () => LinkedAccounts.show(context),
             title: const Text('Linked Accounts'),
+          ),
+          ListTile(
+            onTap: () => PrivacyAndSecuritySettings.show(context),
+            title: const Text('Privacy & Security'),
           ),
           ListTile(
             onTap: () {