+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [0.0.13]
 ### New Features & Screens
 - Add more Button interactivity by modifying `ButtonStyle`
+- Added a function where user can use local authentication such as fingerprint or face ID using [LocalAuthentication] package
+- Created `LocalAuthenticationService` class to deal with [LocalAuthentication] package
 
 ### Bug Fix & Refactor
 
@@ -130,3 +132,4 @@
 [0.0.3]: https://github.com/heeyunlee/cccc/commit/0a5cdbd76996599dfd01a4c290ab7d0723d1e523
 [0.0.2]: https://github.com/heeyunlee/cccc/commit/96682cce01f88cd05830057fc17f703601fc936a
 [collection]: https://pub.dev/packages/collection
+[LocalAuthentication]: https://pub.dev/packages/local_auth